Fix todos datatable unstable columns leading to re-triggering diff animations.
