+++ front/components/assistant/conversation/space/conversations/project_todos/SuggestedTodoItem.tsx
@@ -14,7 +14,7 @@ import {
   TypingAnimation,
   XMarkIcon,
 } from "@dust-tt/sparkle";
-import { memo, useRef, useState } from "react";
+import { memo, useEffect, useRef, useState } from "react";
 
 export interface SuggestedTodoItemProps {
   todo: ProjectTodoType;
@@ -44,9 +44,31 @@ export const SuggestedTodoItem = memo(function SuggestedTodoItem({
     "approve" | "reject" | null
   >(null);
   const measureRef = useRef<HTMLSpanElement>(null);
+  const typingAnimationLockedTextRef = useRef<string | null>(null);
 
   const displayText = stripNewlines(todo.text);
   const showTypingAnimation = isNew && !typingDismissed;
+
+  useEffect(() => {
+    if (!isNew) {
+      typingAnimationLockedTextRef.current = null;
+    }
+  }, [isNew]);
+
+  useEffect(() => {
+    if (!showTypingAnimation) {
+      typingAnimationLockedTextRef.current = null;
+    }
+  }, [showTypingAnimation]);
+
+  let typingAnimationSourceText = displayText;
+  if (showTypingAnimation) {
+    if (typingAnimationLockedTextRef.current === null) {
+      typingAnimationLockedTextRef.current = displayText;
+    }
+    typingAnimationSourceText = typingAnimationLockedTextRef.current;
+  }
+
   const showInProgressTextAnimation = todo.status === "in_progress";
 
   const canAct = viewerUserId !== null && !isReadOnly;
@@ -77,7 +99,7 @@ export const SuggestedTodoItem = memo(function SuggestedTodoItem({
                 aria-hidden
                 className="invisible block w-full min-w-0 break-words text-pretty text-base leading-6"
               >
-                {displayText}
+                {typingAnimationSourceText}
               </span>
             )}
             <TodoMetadataTooltip todo={todo} agentNameById={agentNameById}>
@@ -90,7 +112,7 @@ export const SuggestedTodoItem = memo(function SuggestedTodoItem({
               >
                 {showTypingAnimation ? (
                   <TypingAnimation
-                    text={displayText}
+                    text={typingAnimationSourceText}
                     duration={16}
                     onComplete={() => setTypingDismissed(true)}
                   />
